move state mutability tests into tests/ dir
